// inventory/src/app/superadmin/models/user.model.ts
export interface User {
  _id: string;
  companyName: string;
  companyType: string;
  companyLocation: string;
  companyContact: string;
  CompanyEmail: string;
  companyWebsite: string;
  companyRegNumber: string;
  companyPanNumber: string;
  companyEstd: string;
  companyFocalPersonName: string;
  companyFocalPersonLocation: string;
  companyFocalPersonContact: string;
  companyFocalPersonEmail: string;
  status: boolean;
}